% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{did_estimate_weighted}
\alias{did_estimate_weighted}
\title{Weighted diff-in-diff estimate.
Takes all the same parameters as did_estimate, plus treated.weights and period.weights.}
\usage{
did_estimate_weighted(
  Y,
  N0,
  T0,
  treated.weights = NULL,
  period.weights = NULL,
  ...
)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{treated.weights}{A vector of weights for treated units (length N1).}

\item{period.weights}{A vector of weights for post-treatment periods (length T1).}

\item{...}{additional options for synthdid_estimate_weighted}
}
\value{
an object like that returned by synthdid_estimate_weighted
}
\description{
Weighted diff-in-diff estimate.
Takes all the same parameters as did_estimate, plus treated.weights and period.weights.
}
