% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.synthdid_estimate_weighted}
\alias{format.synthdid_estimate_weighted}
\title{Format a weighted synthdid object}
\usage{
\method{format}{synthdid_estimate_weighted}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Format a weighted synthdid object
}
