% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.synthdid_estimate_weighted}
\alias{plot.synthdid_estimate_weighted}
\title{Plot a weighted synthdid object}
\usage{
\method{plot}{synthdid_estimate_weighted}(x, ...)
}
\arguments{
\item{x}{The object to plot}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Plot a weighted synthdid object
}
