% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.synthdid_event_study}
\alias{plot.synthdid_event_study}
\title{Plot a synthdid_event_study object}
\usage{
\method{plot}{synthdid_event_study}(x, ...)
}
\arguments{
\item{x}{The object to plot, as returned by synthdid_event_study()}

\item{...}{Additional arguments passed to plot_event_study.}
}
\description{
Plot a synthdid_event_study object
}
