% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.synthdid_estimate_weighted}
\alias{print.synthdid_estimate_weighted}
\title{Print a weighted synthdid object}
\usage{
\method{print}{synthdid_estimate_weighted}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a weighted synthdid object
}
