% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{sc_estimate_weighted}
\alias{sc_estimate_weighted}
\title{Weighted synthetic control estimate.
Takes all the same parameters as sc_estimate, plus treated.weights and period.weights.}
\usage{
sc_estimate_weighted(
  Y,
  N0,
  T0,
  treated.weights = NULL,
  period.weights = NULL,
  eta.omega = 1e-06,
  ...
)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{treated.weights}{A vector of weights for treated units (length N1).}

\item{period.weights}{A vector of weights for post-treatment periods (length T1).}

\item{eta.omega}{determines the level of ridge regularization.}

\item{...}{additional options for synthdid_estimate_weighted}
}
\value{
an object like that returned by synthdid_estimate_weighted
}
\description{
Weighted synthetic control estimate.
Takes all the same parameters as sc_estimate, plus treated.weights and period.weights.
}
