% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{synthdid_controls_weighted}
\alias{synthdid_controls_weighted}
\title{Outputs a table of important synthetic controls and their corresponding weights (weighted version)}
\usage{
synthdid_controls_weighted(
  estimates,
  sort.by = 1,
  mass = 0.9,
  weight.type = "omega"
)
}
\arguments{
\item{estimates, }{a list of weighted estimates. Or a single estimate.}

\item{sort.by, }{the index of the estimate to sort by. Defaults to 1.}

\item{mass, }{which controls the length of the table. Defaults to 0.9.}

\item{weight.type, }{'omega' for units, 'lambda' for time periods, 'treated' for treated unit weights}
}
\description{
Outputs a table of important synthetic controls and their corresponding weights (weighted version)
}
