% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_effect_curve_weighted}
\alias{synthdid_effect_curve_weighted}
\title{Outputs the effect curve that was averaged to produce the weighted estimate}
\usage{
synthdid_effect_curve_weighted(estimate)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate_weighted}
}
\description{
Outputs the effect curve that was averaged to produce the weighted estimate
}
