% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{synthdid_placebo_plot_weighted}
\alias{synthdid_placebo_plot_weighted}
\title{For weighted estimator and a placebo, plots treated and synthetic control trajectories.
Requires ggplot2}
\usage{
synthdid_placebo_plot_weighted(
  estimate,
  overlay = FALSE,
  treated.fraction = NULL
)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate_weighted.}

\item{overlay, }{binary, indicates whether plots should be overlaid or shown in different facets. Defaults to FALSE.}

\item{treated.fraction}{as in synthdid_placebo_weighted}
}
\description{
For weighted estimator and a placebo, plots treated and synthetic control trajectories.
Requires ggplot2
}
