% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{synthdid_placebo_weighted}
\alias{synthdid_placebo_weighted}
\title{Computes a placebo variant for the weighted estimator using pre-treatment data only}
\usage{
synthdid_placebo_weighted(estimate, treated.fraction = NULL)
}
\arguments{
\item{estimate, }{as output by synthdid_estimate_weighted}

\item{treated.fraction, }{the fraction of pre-treatment data to use as a placebo treatment period
Defaults to NULL, which indicates that it should be the fraction of post-treatment to pre-treatment data}
}
\description{
Computes a placebo variant for the weighted estimator using pre-treatment data only
}
