% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{synthdid_plot_weighted}
\alias{synthdid_plot_weighted}
\title{Plots treated and synthetic control trajectories for weighted estimates.
Similar to synthdid_plot but accounts for user-specified treated.weights and period.weights.}
\usage{
synthdid_plot_weighted(
  estimates,
  treated.name = "treated (weighted)",
  control.name = "synthetic control",
  ...
)
}
\arguments{
\item{estimates, }{a list of weighted estimates output by synthdid_estimate_weighted. Or a single estimate.}

\item{treated.name, }{the name of the treated curve that appears in the legend. Defaults to 'treated'}

\item{control.name, }{the name of the control curve that appears in the legend. Defaults to 'synthetic control'}

\item{...}{Additional arguments passed to synthdid_plot}
}
\description{
Plots treated and synthetic control trajectories for weighted estimates.
Similar to synthdid_plot but accounts for user-specified treated.weights and period.weights.
}
