% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{synthdid_se_weighted}
\alias{synthdid_se_weighted}
\title{Calculate the standard error of a weighted synthetic diff in diff estimate}
\usage{
synthdid_se_weighted(...)
}
\arguments{
\item{...}{Any valid arguments for vcov.synthdid_estimate_weighted}
}
\description{
Calculate the standard error of a weighted synthetic diff in diff estimate
}
